# src/integration-test.disabled/resources/api/openapi.yaml
openapi: 3.1.0
info:
  title: rtj-management
  contact: {}
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: "1.0"
servers:
- url: http://localhost:59560
  description: Generated server url
...
paths:
...
  /{municipalityId}/{namespace}/errands/{errandId}/stakeholders:
    get:
      tags:
      - Errand stakeholders
      summary: List stakeholders
      operationId: readStakeholders
      parameters:
      - name: municipalityId
        in: path
        required: true
        schema:
          type: string
        example: 2281
      - name: namespace
        in: path
        required: true
        schema:
          type: string
          pattern: "[\\w|\\-]+"
        example: MY_NAMESPACE
      - name: errandId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Stakeholder"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    post:
      tags:
      - Errand stakeholders
      summary: Create stakeholder
      description: Creates a new stakeholder on the errand
      operationId: createStakeholder
      parameters:
      - name: municipalityId
        in: path
        required: true
        schema:
          type: string
        example: 2281
      - name: namespace
        in: path
        required: true
        schema:
          type: string
          pattern: "[\\w|\\-]+"
        example: MY_NAMESPACE
      - name: errandId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Stakeholder"
        required: true
      responses:
        "201":
          description: Successful operation
          headers:
            Location:
              style: simple
              schema:
                type: string
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /{municipalityId}/{namespace}/errands/{errandId}/process-messages:
    post:
      tags:
      - Process messages
      summary: Send a BPMN message to the running process for an errand
      description: "Correlates a BPMN message (e.g. a handläggare's APPROVE/REJECT\
        \ decision, an external event) to the process instance whose business key\
        \ equals this errand id. The matching receive task or message catch event\
        \ resumes with the supplied process variables."
      operationId: sendProcessMessage
      parameters:
      - name: municipalityId
        in: path
        required: true
        schema:
          type: string
        example: 2281
      - name: namespace
        in: path
        required: true
        schema:
          type: string
          pattern: "[\\w|\\-]+"
        example: MY_NAMESPACE
      - name: errandId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProcessMessageRequest"
        required: true
      responses:
        "204":
          description: Message correlated
        "404":
          description: No process instance is waiting for this message on this errand
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
  /{municipalityId}/{namespace}/errands/{errandId}/stakeholders/{stakeholderId}:
    get:
      tags:
      - Errand stakeholders
      summary: Read stakeholder
      operationId: readStakeholder
      parameters:
      - name: municipalityId
        in: path
        required: true
        schema:
          type: string
        example: 2281
      - name: namespace
        in: path
        required: true
        schema:
          type: string
          pattern: "[\\w|\\-]+"
        example: MY_NAMESPACE
      - name: errandId
        in: path
        required: true
        schema:
          type: string
      - name: stakeholderId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stakeholder"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    delete:
      tags:
      - Errand stakeholders
      summary: Delete stakeholder
      operationId: deleteStakeholder
      parameters:
      - name: municipalityId
        in: path
        required: true
        schema:
          type: string
        example: 2281
      - name: namespace
        in: path
        required: true
        schema:
          type: string
          pattern: "[\\w|\\-]+"
        example: MY_NAMESPACE
      - name: errandId
        in: path
        required: true
        schema:
          type: string
      - name: stakeholderId
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: Successful operation
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    patch:
      tags:
      - Errand stakeholders
      summary: Update stakeholder
      operationId: updateStakeholder
      parameters:
      - name: municipalityId
        in: path
        required: true
        schema:
          type: string
        example: 2281
      - name: namespace
        in: path
        required: true
        schema:
          type: string
          pattern: "[\\w|\\-]+"
        example: MY_NAMESPACE
      - name: errandId
        in: path
        required: true
        schema:
          type: string
      - name: stakeholderId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Stakeholder"
        required: true
      responses:
        "204":
          description: Successful operation
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
components:
  schemas:
    Problem:
      type: object
      properties:
        detail:
          type: string
        title:
          type: string
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        status:
          type: integer
          format: int32
    ConstraintViolationProblem:
      type: object
      properties:
        type:
          type: string
          format: uri
        status:
          type: integer
          format: int32
        violations:
          type: array
          items:
            $ref: "#/components/schemas/Violation"
        title:
          type: string
        detail:
          type: string
        causeAsProblem:
          $ref: "#/components/schemas/ThrowableProblem"
        instance:
          type: string
          format: uri
    ThrowableProblem:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        causeAsProblem: {}
    Violation:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
...
    ContactChannel:
      type: object
      description: ContactChannel model
      properties:
        key:
          type: string
          description: The key of the contact channel
          examples:
          - Email
        value:
          type: string
          description: The value of the contact channel
          examples:
          - joe.doe@example.com
    Stakeholder:
      type: object
      description: Stakeholder
      properties:
        id:
          type: string
          description: Unique identifier
          examples:
          - cb20c51f-fcf3-42c0-b613-de563634a8ec
          readOnly: true
        externalId:
          type: string
          description: External id for the stakeholder
          examples:
          - 81471222-5798-11e9-ae24-57fa13b361e1
        externalIdType:
          type: string
          description: Type of external id
          examples:
          - PRIVATE
        role:
          type: string
          description: Role of the stakeholder — validated against StakeholderRoleRegistry
            for the errand's typeSlug
          examples:
          - FOSTER_PARENT
        firstName:
          type: string
          description: First name
          examples:
          - Joe
        lastName:
          type: string
          description: Last name
          examples:
          - Doe
        organizationName:
          type: string
          description: Organization name
          examples:
          - Sundsvalls kommun
        address:
          type: string
          description: Address
          examples:
          - Storgatan 1
        careOf:
          type: string
          description: Care of
          examples:
          - c/o Doe
        zipCode:
          type: string
          description: Zip code
          examples:
          - "85248"
        city:
          type: string
          description: City
          examples:
          - Sundsvall
        country:
          type: string
          description: Country
          examples:
          - Sweden
        contactChannels:
          type: array
          description: Contact channels for the stakeholder
          items:
            $ref: "#/components/schemas/ContactChannel"
    ProcessMessageRequest:
      type: object
      description: "Request body for correlating a BPMN message to the process instance\
        \ currently running for an errand. The errand id is used as the process business\
        \ key, so the message is delivered to that specific process. Use this whenever\
        \ something outside the process (a handläggare action, an external event,\
        \ an admin override) needs to resume or interact with a running process instance."
      properties:
        messageName:
          type: string
          description: "BPMN message name, matching the `name` attribute on the `<bpmn:message>`\
            \ element the receive task references"
          examples:
          - PaymentDecisionReceived
          minLength: 1
        variables:
          type: object
          additionalProperties: {}
          description: Process variables to set when correlating the message
      required:
      - messageName
...
  securitySchemes: {}